Renaming bot_orm to orm, and fixing an import issue in some tests
