--- inhouse_bot/bot_orm/__init__.py
@@ -1,7 +1,0 @@
-from inhouse_bot.bot_orm.session.session_handler import session_scope, get_session, bot_declarative_base
-
-from inhouse_bot.bot_orm.tables.game import Game
-from inhouse_bot.bot_orm.tables.game_participant import GameParticipant
-from inhouse_bot.bot_orm.tables.player import Player
-from inhouse_bot.bot_orm.tables.player_rating import PlayerRating
-from inhouse_bot.bot_orm.tables.queue_player import QueuePlayer
