--- inhouse_bot/bot_orm/tables/__init__.py
@@ -1,1 +1,0 @@
-from inhouse_bot.bot_orm.tables.player_rating import PlayerRating
