+++ inhouse_bot/matchmaking_logic/evaluate_game.py
@@ -3,7 +3,7 @@ import math
 
 import trueskill
 
-from inhouse_bot.bot_orm import Game
+from inhouse_bot.orm import Game
 
 
 def evaluate_game(game: Game) -> float:
