+++ inhouse_bot/orm/__init__.py
@@ -1,0 +1,7 @@
+from inhouse_bot.orm.session.session_handler import session_scope, get_session, bot_declarative_base
+
+from inhouse_bot.orm.tables.game import Game
+from inhouse_bot.orm.tables.game_participant import GameParticipant
+from inhouse_bot.orm.tables.player import Player
+from inhouse_bot.orm.tables.player_rating import PlayerRating
+from inhouse_bot.orm.tables.queue_player import QueuePlayer
