+++ inhouse_bot/orm/tables/__init__.py
@@ -1,0 +1,1 @@
+from inhouse_bot.orm.tables.player_rating import PlayerRating
