+++ tests/test_cog.py
@@ -1,7 +1,7 @@
 from discord.ext import commands
 from discord.ext.commands import group
 
-from inhouse_bot.bot_orm import session_scope
+from inhouse_bot.orm import session_scope
 from inhouse_bot.cogs.cogs_utils.validation_dialog import checkmark_validation
 from inhouse_bot.common_utils.fields import roles_list
 from inhouse_bot.common_utils.get_last_game import get_last_game
